refactor(core): move namespace status out of full materialization
