+++ crates/loonfs-core/src/namespace/full_materialization.rs
@@ -1,12 +1,9 @@
 use crate::checkpoint::{
-    head_from_manifest, load_namespace_manifest_envelope, load_verified_manifest_materialization,
-    ManifestLoadError,
+    head_from_manifest, load_verified_manifest_materialization, ManifestLoadError,
 };
-use crate::error::CoreError;
 use crate::metadata::MetadataState;
 use crate::namespace::catalog::{
-    load_namespace_catalog_entry, namespace_initialization_state, NamespaceCatalogLoadError,
-    NamespaceInitializationError, NamespaceInitializationState, VerifiedNamespaceCatalogEntry,
+    load_namespace_catalog_entry, NamespaceCatalogLoadError, VerifiedNamespaceCatalogEntry,
 };
 use crate::namespace::control::{read_head_object, read_lease_object, ControlObjectLoadError};
 use crate::wal::{
@@ -14,11 +11,8 @@ use crate::wal::{
     WalChainLoadRequest, WalReplayError,
 };
 use loonfs_api::wire::control::{HeadState, LeaseState, NamespaceDescriptorState, NamespaceState};
-use loonfs_api::{wal_segment_id_start_seq, ChangeSeq, ContentStoreId, ManifestId, NamespaceId};
-use loonfs_objectstore::{
-    keys::{namespace_descriptor, namespace_head, wal_segment_id_from_key, wal_segment_prefix},
-    ObjectStore,
-};
+use loonfs_api::{ChangeSeq, ContentStoreId, NamespaceId};
+use loonfs_objectstore::{keys::namespace_head, ObjectStore};
 use serde::{Deserialize, Serialize};
 use std::mem::size_of;
 use thiserror::Error;
@@ -133,31 +127,6 @@ fn wal_tip_decoded_bytes(pointer: Option<&loonfs_api::wire::control::WalSegmentP
                 + pointer.payload_checksum.len()
         })
         .unwrap_or(0)
-}
-
-/// Lightweight namespace head status.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NamespaceHeadSummary {
-    pub namespace_id: NamespaceId,
-    pub head_seq: ChangeSeq,
-    pub current_manifest_id: Option<ManifestId>,
-    pub latest_checkpoint_id: Option<String>,
-    /// WAL segment objects positioned past the loaded manifest.
-    ///
-    /// Derived from position-ordered object names, not from walking the
-    /// chain: an inspection count for maintenance gating and operators, not
-    /// a validated chain length.
-    pub wal_tail_segments: u64,
-    pub retention_floor_seq: ChangeSeq,
-}
-
-/// Opaque ETag probe for the namespace head object.
-///
-/// This only proves that the durable head object identity still matches a
-/// previously reconstructed materialization.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NamespaceHeadEtagProbe {
-    pub head_etag: String,
 }
 
 /// Error while reconstructing a full namespace materialization.
@@ -460,153 +429,6 @@ async fn ensure_head_etag_still_current<S: ObjectStore + ?Sized>(
     Ok(())
 }
 
-pub async fn load_namespace_head_summary<S: ObjectStore + ?Sized>(
-    store: &S,
-    expected_namespace: &NamespaceId,
-) -> Result<NamespaceHeadSummary, CoreError> {
-    match namespace_initialization_state(store, expected_namespace).await {
-        Ok(NamespaceInitializationState::Complete) => {}
-        Ok(NamespaceInitializationState::Absent) => {
-            return Err(CoreError::FullMaterialization(
-                FullMaterializationLoadError::LoadNamespaceDescriptor(
-                    ControlObjectLoadError::MissingObject {
-                        object_key: namespace_descriptor(expected_namespace.as_str()),
-                    },
-                ),
-            ));
-        }
-        Ok(NamespaceInitializationState::Partial) => {
-            return Err(CoreError::NamespacePartiallyInitialized {
-                namespace_id: expected_namespace.clone(),
-            });
-        }
-        Err(error) => return Err(map_namespace_initialization_error_to_core(error)),
-    }
-
-    let loaded_head = read_head_object(store, expected_namespace)
-        .await
-        .map_err(|error| {
-            CoreError::FullMaterialization(FullMaterializationLoadError::LoadHead(error))
-        })?;
-    if loaded_head.envelope.state.state == NamespaceState::Deleted {
-        return Err(CoreError::NamespaceDeleted {
-            namespace_id: expected_namespace.clone(),
-        });
-    }
-    let head = loaded_head.envelope.state;
-    let manifest_id = head.current_manifest_id.ok_or_else(|| {
-        CoreError::FullMaterialization(FullMaterializationLoadError::MissingCurrentManifest {
-            namespace_id: expected_namespace.clone(),
-        })
-    })?;
-    let manifest_materialization_seq =
-        load_namespace_manifest_envelope(store, expected_namespace, manifest_id)
-            .await
-            .map_err(|error| {
-                CoreError::FullMaterialization(FullMaterializationLoadError::ManifestLoad(error))
-            })?
-            .payload
-            .head_seq;
-    let wal_tail_segments = if head.visible_wal_tip.is_some() {
-        count_wal_tail_segments_by_position(store, expected_namespace, manifest_materialization_seq)
-            .await?
-    } else {
-        0
-    };
-    Ok(NamespaceHeadSummary {
-        namespace_id: head.namespace_id,
-        head_seq: head.seq,
-        current_manifest_id: head.current_manifest_id,
-        latest_checkpoint_id: head.latest_checkpoint_id,
-        wal_tail_segments,
-        retention_floor_seq: head.retention_floor_seq,
-    })
-}
-
-/// Counts WAL tail segments from their position-ordered object names.
-///
-/// Status is an inspection surface, so the count comes from one listing
-/// instead of loading and validating segment bodies: segment file names
-/// carry their `start_seq`, and every chain segment past the manifest starts
-/// above it. Objects that lost a head race are counted until reclamation
-/// removes them, which can only over-trigger maintenance, never starve it.
-/// Recovery authority stays with the head and chain.
-async fn count_wal_tail_segments_by_position<S: ObjectStore + ?Sized>(
-    store: &S,
-    namespace_id: &NamespaceId,
-    manifest_materialization_seq: ChangeSeq,
-) -> Result<u64, CoreError> {
-    let keys = store
-        .list_prefix(&wal_segment_prefix(namespace_id.as_str()))
-        .await
-        .map_err(|error| CoreError::Store(format!("list WAL tail segments: {error}")))?;
-    let tail_segments = keys
-        .iter()
-        .filter_map(|key| wal_segment_id_from_key(key))
-        .filter_map(wal_segment_id_start_seq)
-        .filter(|start_seq| *start_seq > manifest_materialization_seq)
-        .count();
-    u64::try_from(tail_segments)
-        .map_err(|_| CoreError::Store("WAL tail segment count overflow".to_owned()))
-}
-
-#[tracing::instrument(
-    level = "info",
-    name = "loon.phase",
-    err,
-    skip_all,
-    fields(phase = "probe_namespace_head_etag", key_class = "namespace_head")
-)]
-pub async fn probe_namespace_head_etag<S: ObjectStore + ?Sized>(
-    store: &S,
-    expected_namespace: &NamespaceId,
-) -> Result<NamespaceHeadEtagProbe, CoreError> {
-    NamespaceId::parse(expected_namespace.as_str())?;
-    let object_key = namespace_head(expected_namespace.as_str());
-    let metadata = store
-        .head(&object_key)
-        .await
-        .map_err(|err| {
-            CoreError::FullMaterialization(FullMaterializationLoadError::LoadHead(
-                ControlObjectLoadError::Store(err.to_string()),
-            ))
-        })?
-        .ok_or_else(|| {
-            CoreError::FullMaterialization(FullMaterializationLoadError::LoadHead(
-                ControlObjectLoadError::MissingObject {
-                    object_key: object_key.clone(),
-                },
-            ))
-        })?;
-    let head_etag = metadata.etag.ok_or(CoreError::FullMaterialization(
-        FullMaterializationLoadError::MissingHeadEtag { object_key },
-    ))?;
-    Ok(NamespaceHeadEtagProbe { head_etag })
-}
-
-fn map_namespace_initialization_error_to_core(error: NamespaceInitializationError) -> CoreError {
-    match error {
-        NamespaceInitializationError::InvalidNamespaceId(error) => {
-            CoreError::InvalidNamespaceId(error)
-        }
-        NamespaceInitializationError::LoadNamespaceDescriptor(error) => {
-            CoreError::FullMaterialization(FullMaterializationLoadError::LoadNamespaceDescriptor(
-                error,
-            ))
-        }
-        NamespaceInitializationError::LoadContentStoreDescriptor(error) => {
-            CoreError::FullMaterialization(
-                FullMaterializationLoadError::LoadContentStoreDescriptor(error),
-            )
-        }
-        NamespaceInitializationError::InspectNamespaceDescriptor(_)
-        | NamespaceInitializationError::InspectNamespaceHead(_)
-        | NamespaceInitializationError::InspectNamespaceLease(_) => {
-            CoreError::Store(error.to_string())
-        }
-    }
-}
-
 fn ensure_reconstructed_head_matches(
     current_head: &HeadState,
     reconstructed: &HeadState,
@@ -636,6 +458,7 @@ fn ensure_reconstructed_head_matches(
 mod tests {
     use super::*;
     use crate::checkpoint::load_verified_manifest_materialization;
+    use crate::namespace::status::load_namespace_head_summary;
     use crate::{BootstrapOptions, NamespaceEngine, ReadOptions, WriteOptions};
     use async_trait::async_trait;
     use bytes::Bytes;
