+++ crates/loonfs-core/src/namespace/mod.rs
@@ -11,6 +11,7 @@ pub(crate) mod delete;
 pub(crate) mod fork;
 pub(crate) mod full_materialization;
 pub(crate) mod lease;
+pub(crate) mod status;
 
 pub use bootstrap::BootstrapNamespaceError;
 pub use loonfs_api::wire::control::{HeadState, HeadStateEnvelope, LeaseState, LeaseStateEnvelope};
